Speeding up the creation of a pool of lb members
With a large number of load balancer members, creating and updating
could take a long time, for two listeners of 150 members it took about
an hour of time.
Now all members are added with a single request, which reduces the time
to about 10 minutes for the above number of members.